+++ src/ESPAsync_WiFiManager-Impl.h
@@ -14,7 +14,7 @@
   Built by Khoi Hoang https://github.com/khoih-prog/ESPAsync_WiFiManager
   Licensed under MIT license
   
-  Version: 1.12.0
+  Version: 1.12.1
 
   Version Modified By  Date      Comments
   ------- -----------  ---------- -----------
@@ -50,6 +50,7 @@
   1.10.0  K Hoang      29/12/2021 Fix `multiple-definitions` linker error and weird bug related to src_cpp
   1.11.0  K Hoang      17/01/2022 Enable compatibility with old code to include only ESP_WiFiManager.h
   1.12.0  K Hoang      10/02/2022 Add support to new ESP32-S3
+  1.12.1  K Hoang      11/02/2022 Add LittleFS support to ESP32-C3. Use core LittleFS instead of Lorol's LITTLEFS for v2.0.0+
  *****************************************************************************************************************************/
 
 #pragma once